pyproject.toml
[project]
name = "continual-topic-model"
version = "0.1.0"
description = "Continual Topic Modeling with VAE + LLM-guided Memory Management"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.0",
    "numpy>=1.24",
    "scipy>=1.10",
    "scikit-learn>=1.3",
    "spacy>=3.7",
    "pandas>=2.0",
    "tqdm",
    "pyyaml",
    "sentence-transformers>=2.2",
    "google-genai>=1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
